Component splits, Added Meta.
