+++ mycomponents/chakra-next-image.tsx
